Clarify the meaning of remote commands.
You are only ever talking to your local copy of the remote, except for
the new command I added an example of, so I tried to make that clear.

Also fixed a nit with switching branches: you *must* run make afterward
(at least you do if you expect to test anything).  And added a reference to
git worktree...I'm fine if that gets rejected; I haven't used it yet
myself :)