Update the Org Repository Policy per the 2022 PSC.
This is a result of the discussion started by the 2022 Python Steering Council
on https://discuss.python.org/t/new-python-organization-repository-policy/17376
and
https://mail.python.org/archives/list/python-dev@python.org/message/FAASOMUNQZDJBCOTZK6I2KH55SDOONF6/.

Petr's text from the discuss thread is used, omitting any mention of the PSF
Contributor Agreement (CLA) as discussion revealed that it may not make sense
for all repos (see thread).

It also adopts a suggestion from Joannah to have people discuss additions in
our Committers forum before bringing an addition to the Python Steering Council.