Remove myself as an Android expert
I'm not an expert; the first time I've installed the Android SDK
was to review Android/README.md when support was being finalzed.
(FWIW, I like to think that a beginner's point of view helped there.)

As a PEP sponsor, I helped with the *process* side of getting the
changes in.

I'm happy to help with Android support, as with any other part of
Python; it's not special to me, nor I to it.