Change reference in GC design doc
Prior to the change, the link led to a fixed version 3.8. Currently, it's not the latest version.

As the time passes by, it's going to be outdated further and further.

It might be the best solution to refer to the latest dev version of the doc,
this way we can ensure that reader will access the latest data.