contrib: restructure Getting Started - Setup and Building into a Workflows section (#1467)
* todo: why duplicate toctree docs here?

* general re-organization

* contrib: replace Get Started with Workflows

* a central links file

This makes it easier to move content around without worrying about where
the links are defined.  It will also be easier to have canonical links
that can be used from multiple places in the guide.

* contrib: restructure Getting Started - Setup and Building into a Workflows section

The Getting Started section has a lot of technical detail.  I thought it
would be better moved to the end, and referenced from the contribution
paths earlier in the guide.  I haven't updated the paths yet.

Also:

- I added a central links.rst that can be included anywhere.
- I omitted the "Editors and tools" section. It doesn't provide any useful information.
- Perhaps the Workflows section could be further categorized.
- The `c_blah_start` and `c_blah_end` comments are only for the temporary inclusions being used to demonstrate the new organization.

* re-order the columns